dialog refactor
